fix bug with empty field
